Localization Manager
Added a localization manager class containing the definition for the supported languages and initializing the system. The CSVLoader reads a csv file in order to  return the correct key-value pair. Every key (f.e. clock_description) has an associated value per language (f.e. Time:).
In order to use the localized strings on an object, use TextMeshPro and add the TextLocalizerUI MonoBehaviour. Type in the key that you want and it will update when the game runs.