Removed IHarvest, Updated Plant Logic
Removed IHarvest since Harvests are going to be ripe plants.
Plant logic now follows the Type Object pattern. A plant has a species object containing the species related information with all it's Stages. A stage represents a phase of a plant (from seeding to decaying).
A Stage contains information about the time it takes to reach the next stage and the Need of the plant.
Needs can be reaching a certain amount of humidity/fertilizer etc...