Insects actually flee
In the end i decided to go for a state-machine using an enum.
The reason it took so long was that the script was written without absolutely any thought, so everything was a mess. It took me hours to fix a few minor bugs. A lot of complicated systems act together (pathfinding, hazard spawning, insect behaviour) and none of them were properly planned out.